Add UI Automator and some tests with it
Separate AndroidManifest.xml in "androidTest" folder was added to
override UI Automator's minSdk requirements.

Clicking with UI Automator on a multi-word link located at the end of
long string turned out to be unreliable. It clicks "at the center of the
visible bounds of the UI element", which can be a blank space if the
link is word-wrapped. Fixes by adding line break to about.htm.